handle resetting the timer when loading a map
