+++ src/data/state.js
@@ -4,6 +4,7 @@ import {
 	TILE_SIZE,
 } from './grid'
 import { createObservable } from '../helpers/createObservable'
+import { resetTimer } from '../helpers/timer'
 import * as maps from '../maps'
 
 
@@ -12,9 +13,10 @@ import * as maps from '../maps'
 
 let stateObject = {
 	currentTile: 0,
-	lastTimerUpdate: 0,
-	frame: 0,
 	entities: [],
+	frame: 0,
+	isVictory: 0,
+	lastTimerUpdate: 0,
 	map: null,
 	paused: 1,
 	placeX: 0,
@@ -23,15 +25,21 @@ let stateObject = {
 		timeBonus: 0,
 	},
 	timeRemaining: 0,
-	victory: 0,
 }
 
 export let state = createObservable(new Proxy(stateObject, {
 	set(target, key, value) {
 		if (key === 'map') {
-			let map = maps[value];
+			let map = maps[value]
+			resetTimer()
 			map.reset()
-			target[key] = map;
+			target.currentTile = 0
+			target.isVictory = 0
+			target.entities = map.objects
+			target.placeX = 0
+			target.placeY = 0
+			target.timeRemaining = map.delay
+			target.map = map
 			return 1
 		}
